# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FormaOS API",
    "description": "FormaOS is a compliance management platform. This API provides programmatic access to organizations, tasks, evidence, compliance metrics, controls, members, webhooks, integrations, API keys, frameworks, AI chat, reports, notifications, data residency, SOC 2, certificates, and audit logs.\n\nAll API keys use the `fos_` prefix. Include the key as a Bearer token in the Authorization header.",
    "version": "1.0.0",
    "contact": {
      "name": "FormaOS API Support",
      "email": "Formaos.team@gmail.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://formaos.com",
      "description": "Production"
    },
    {
      "url": "https://staging.formaos.com",
      "description": "Staging"
    }
  ],
  "security": [
    {
      "apiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/data-residency": {
      "get": {
        "summary": "Get data residency settings",
        "description": "Returns the organization's current data residency region and all available regions.",
        "operationId": "getDataResidency",
        "tags": ["Data Residency"],
        "responses": {
          "200": {
            "description": "Data residency settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "currentRegion": {
                      "type": "string",
                      "enum": ["au", "us", "eu"]
                    },
                    "availableRegions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "name": { "type": "string" },
                          "description": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "patch": {
        "summary": "Update data residency region",
        "description": "Changes the organization's data residency region. Requires admin role.",
        "operationId": "updateDataResidency",
        "tags": ["Data Residency"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["region"],
                "properties": {
                  "region": {
                    "type": "string",
                    "enum": ["au", "us", "eu"],
                    "description": "Target data residency region"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Region updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "region": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "fos_*",
        "description": "FormaOS API key with `fos_` prefix. Pass as `Authorization: Bearer fos_...`"
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentication required -- missing or invalid Bearer token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions for this operation",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "BadRequest": {
        "description": "Invalid request body or parameters",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error code or message"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error description"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        },
        "required": ["error"]
      },
...
    }
  }
}